requirements.txt
fastapi
pydantic
langchain~=0.0.184
python-dotenv
openai
markdown
uvicorn
uvloop
bs4
unstructured
chromadb
tiktoken
pdfminer-six
tabulate
psycopg
psycopg2-binary
faiss-cpu
duckduckgo-search
pdf2image
typing
python-multipart